requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
appdirs==1.4.4
appnope==0.1.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
docker-pycreds==0.4.0
executing==2.0.1
filelock==3.13.0
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
h5py==3.10.0
idna==3.4
ipykernel==6.26.0
ipython==8.17.1
ipywidgets==8.1.1
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.5.0
jupyter_core==5.5.0
jupyterlab-widgets==3.0.9
lightning==2.1.0
lightning-utilities==0.9.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
nest-asyncio==1.5.8
networkx==3.2.1
numpy==1.26.1
packaging==23.2
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
python-dateutil==2.8.2
pytorch-lightning==2.1.0
PyYAML==6.0.1
pyzmq==25.1.1
requests==2.31.0
rich==13.6.0
sentry-sdk==1.32.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
sympy==1.12
torch==2.1.0
torchmetrics==1.2.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
typing_extensions==4.8.0
urllib3==2.0.7
wandb==0.15.12
wcwidth==0.2.8
widgetsnbextension==4.0.9
yarl==1.9.2
